Main

Description

Este ficheiro contém a função principal do jogo e todos os mecanismos usados.

Documentation

data Manager

Constructors

Manager

Fields

state :: State

pid :: Int

step :: Int

before :: Integer

delta :: Integer

delay :: Integer

loadManager

Arguments

:: Manager
Manager

Função que lê a data Manager.

loadManager2

Arguments

:: Maze
Labirinto

-> Manager
Manager

Função que lê a data Manager mas que recebe um labirinto.

generatePlayerList

Arguments

:: Maze
Labirinto

-> [Player]
Lista de jogadores

Função que gera uma lista de três jogadores.

updateControlledPlayer

Arguments

:: Key
Tecla

-> Manager
Manager atual

-> Manager
Manager novo

Função que controla o input do jogador.

updateScreen

Arguments

:: Window
Window atual

-> ColorID
ID da cor

-> Manager
Manager atual

-> Curses ()
Window nova

Função que atualiza a Window do nCurses.

currentTime

Arguments

:: IO Integer
Tempo atual

Função que retorna o tempo atual.

updateTime

Arguments

:: Integer
Tempo atual

-> Manager
Manager atual

-> Manager
Manager novo

Função que atualiza o tempo.

resetTimer

Arguments

:: Integer
Tempo atual

-> Manager
Manager atual

-> Manager
Manager novo

Função que dá reset ao tempo.

nextFrame

Arguments

:: Integer
Tempo atual

-> Manager
Manager atual

-> Int
Modo de jogo

-> Int
Pausa

-> Manager
Manager novo

Função que invoca o jogo principal e que aumenta os steps.

nextLevel

Arguments

:: Manager
Manager atual

-> Manager
Manager novo

Função que aumenta o nível do jogo.

nextLevelAUX

Arguments

:: Int
Nível do jogo

-> Manager
Manager atual

-> Manager
Manager novo

Função que invoca a próxima função necessária para completar a passagem de nível.

nextLevelAUX2

Arguments

:: Int
Nível do jogo

-> State
Estado atual

-> State
Estado novo

Função que altera o nível no estado do jogo.

loop

Arguments

:: Int
Modo de jogo

-> Int
Pausa

-> Window
Window atual

-> Manager
Manager atual

-> Curses ()
Window nova

Função que invoca as funções do jogo.

main :: IO ()

Função principal.

mainM

Arguments

:: Maze
Labirinto

-> IO ()

Função principal para a utilização de um labirinto criada pela Tarefa1.

instructions :: IO ()

Função para verificar as definições.

exit :: IO ()

Função para terminar o jogo.