Tarefa1

Description

O objetivo deste módulo é a geração de um labirinto, consoante a largura, altura e semente de geração aleatória dada.

Geração do labirinto completo

generateMaze

Arguments

:: Int
Largura do labirinto

-> Int
Altura do labirinto

-> Int
Semente para a geração aleatória

-> Maze
Labirinto final

Função geradora de um labirinto.

Construção do labirinto inicial

numeroPecas

Arguments

:: Int
Largura do labirinto

-> Int
Altura do labirinto

-> Int
Número total de peças

Função que determina o número total de peças que o labirinto vai ter.

generateRandoms

Arguments

:: Int
Número total de peças

-> Int
Semente para a geração aleatória

-> [Int]
Lista com todos os números aleatórios

Função que gera os números aleatórios para cada coordenada do labirinto. Tornando-o assim aleatório.

dividirListas

Arguments

:: Int
Largura do labirinto

-> [a]
Lista com todos os números aleatórios

-> [[a]]
Matriz

Função que divide uma lista pela largura do labirinto desejado, criando assim uma matriz.

trocarPecas

Arguments

:: [[Int]]
Matriz

-> Maze
Labirinto

Função que troca os números aleatórios pela sua correspondência certa.

trocarPecasAUX

Arguments

:: [Int]
Lista de números aleatórios

-> Corridor
Corredor com as peças bem representadas

Função que auxilia a troca dos números aleatórios pela sua correspondência.

Com as seguintes condições:

número aleatório == 3 representa Food Big

número aleatório >= 70 e <= 99 representa Wall

senão número aleatório representa Food Little

Inserção das paredes

adicionarParedes

Arguments

:: Maze
Labirinto inicial

-> Maze
Labirinto com as paredes à volta

Função que adiciona as paredes ao labirinto.

adicionarParedesCimaBaixo

Arguments

:: Maze
Labirinto

-> Maze
Labirinto

Função que adiciona as paredes no topo e na base do labirinto

adicionarParedesLados

Arguments

:: Maze
Labirinto

-> Maze
Labirinto

Função que adiciona as paredes no lado esquerdo e direito do labirinto

paredeCheia

Arguments

:: Int
Largura do labirinto

-> Corridor
Corredor com apenas paredes

Função auxiliar que cria um corredor com o tamanho da largura do labirinto, com apenas paredes.

Inserção do túnel

inserirVazio

Arguments

:: Maze
Labirinto com as paredes

-> Int
Altura do labirinto

-> Maze
Labirinto com as paredes e o túnel

Função que chama a função certa dependendo se a altura do labirinto é par ou ímpar.

inserirVazioI

Arguments

:: Maze
Labirinto

-> Int
Altura do labirinto dividida por 2

-> Maze
Labirinto

Função que adiciona o túnel nas coordenadas certas num labirinto com altura ímpar.

inserirVazioP

Arguments

:: Maze
Labirinto

-> Int
Altura do labirinto dividida por 2

-> Maze
Labirinto

Função que adiciona o túnel nas coordenadas certas num labirinto com altura par.

Inserção da casa dos fantasmas

casaFantasmas

Arguments

:: Maze
Labirinto com as paredes e o túnel

-> Int
Largura do labirinto

-> Int
Altura do labirinto

-> Maze
Labirinto com as paredes, o túnel e a casa dos fantasmas

Função que chama a função certa dependendo se a largura do labirinto é par ou ímpar.

procurarLarguraCF

Arguments

:: Int
Largura do labirinto

-> Int
Largura do labirinto a deixar até ao início da coordenada pretendida

Função auxiliar que determina, dependendo se a largura do labirinto é par ou ímpar, a largura do labirinto que vai ser preciso deixar até ao início da casa dos fantasmas.

procurarAlturaCF

Arguments

:: Int
Altura do labirinto

-> Int
Altura do labirinto a deixar até ao início do corredor pretendido

Função auxiliar que determina a altura a deixar até ao início da casa dos fantasmas.

criarCasaFPar

Arguments

:: Maze
Labirinto

-> Int
Altura do labirinto a deixar até ao início do corredor pretendido

-> Int
Largura do labirinto a deixar até ao início da coordenada pretendida

-> Int
Varíavel que começa em 0 e incrementa sempre que esta função volta a ser chamada

-> Maze
Labirinto

Função que adiciona ao labirinto, de largura par, a casa dos fantasmas. Recorrendo a todas as funções necessárias para tal feito.

criarCasaFImpar

Arguments

:: Maze
Labirinto

-> Int
Altura do labirinto a deixar até ao início corredor pretendido

-> Int
Largura do labirinto a deixar até ao início da coordenada pretendida

-> Int
Varíavel que começa em 0 e incrementa sempre que esta função volta a ser chamada

-> Maze
Labirinto

Função que adiciona ao labirinto, de largura ímpar, a casa dos fantasmas. Recorrendo a todas as funções necessárias para tal feito.

insereVaziosCF

Arguments

:: Maze
Labirinto

-> Int
Altura do labirinto a deixar até ao início do corredor pretendido

-> Int
Largura do labirinto a deixar até ao início da coordenada pretendida

-> Int
Número de peças vazias que o corredor vai ter

-> Maze
Labirinto

Função que percorre o labirinto até encontrar o corredor pretendido. Chama, depois, a função certa para a inserção do corredor pretendido.

insereVaziosCFAUX

Arguments

:: Corridor
Corredor

-> Int
Largura do labirinto a deixar até ao início da coordenada pretendida

-> Int
Número de peças vazias que o corredor vai ter

-> Corridor
Corredor com o número certo de peças vazias nas coordenadas certas

Função auxiliar que insere o corredor com o número certo de peças vazias, nas coordenadas certas.

insereCFPar

Arguments

:: Maze
Labirinto

-> Int
Altura do labirinto a deixar até ao início do corredor pretendido

-> Int
Largura do labirinto a deixar até ao início da coordenada pretendida

-> Int
Variável que gere qual dos três corredores é para ser construído

-> Maze
Labirinto

Função que chama as funções certas para a inserção dos três corredores da casa dos fantasmas pretendidos, num labirinto com largura par.

insereCFPar1

Arguments

:: Maze
Labirinto

-> Int
Altura do labirinto a deixar até ao início do corredor pretendido

-> Int
Largura do labirinto a deixar até ao início da coordenada pretendida

-> Int
Variável que gere qual dos três corredores é para ser construído

-> Maze
Labirinto

Função que percorre um labirinto até encontrar o corredor pretendido. Chama, depois, a função certa para a inserção dos três corredores pretendidos.

insereCFPar1AUX

Arguments

:: Corridor
Corredor

-> Int
Largura do labirinto a deixar até ao início da coordenada pretendida

-> Int
Variável que gere qual dos três corredores é para ser construído

-> Corridor
Corredor pretendido

Função auxiliar que percorre um corredor até encontrar a coordenada pretendida. Insere, depois, o corredor pretendido da casa dos fantasmas nas coordenadas certas.

insereCFImpar

Arguments

:: Maze
Labirinto

-> Int
Altura do labirinto a deixar até ao início do corredor pretendido

-> Int
Largura do labirinto a deixar até ao início da coordenada pretendida

-> Int
Variável que gere qual dos três corredores é para ser construído

-> Maze
Labirinto

Função que chama as funções certas para a inserção das peças pretendidas da casa dos fantasmas nos corredores certos, num labirinto com largura ímpar.

insereCFImpar1

Arguments

:: Maze
Labirinto

-> Int
Altura do labirinto a deixar até ao início do corredor pretendido

-> Int
Largura do labirinto a deixar até ao início da coordenada pretendida

-> Int
Variável que gere qual dos três corredores é para ser construído

-> Maze
Labirinto

Função que percorre um labirinto até encontrar o corredor pretendido. Chama, depois, a função certa para a inserção das peças pretendidas nos corredores certos.

insereCFImpar1AUX

Arguments

:: Corridor
Corredor

-> Int
Largura do labirinto a deixar até ao início da coordenada pretendida

-> Int
Variável que gere qual dos três corredores é para ser construído

-> Corridor
Corredor pretendido

Função auxiliar que percorre um corredor até encontrar a coordenada pretendida. Insere, depois, as peças pretendidas da casa dos fantasmas nas coordenadas certas.