Tarefa2

Description

O objetivo deste módulo é a realização de uma jogada, consoante a orientação dessa mesma e principalmente o estado do jogo.

Realização da jogada

play

Arguments

:: Play
Jogada pretendida

-> State
Estado inicial

-> State
Estado final

Função principal que realiza a jogada pretendida.

Funções auxiliares

findPlayerState

Arguments

:: State
Estado

-> Int
ID do jogador que queremos

-> PlayerState
Estado do jogador certo

Função que indica qual é o estado do jogador que queremos mover.

findPlayer

Arguments

:: State
Estado

-> Int
ID do jogador que queremos

-> Player
Jogador pretendido

Função que retorna o jogador pretendido.

findPlayerAUX

Arguments

:: Int
ID do jogador que queremos

-> [Player]
Lista de todos os jogadores

-> Player
Jogador pretendido

Função auxiliar que encontra o jogador que queremos mover.

findPacmanPlayer

Arguments

:: [Player]
Lista de todos os jogadores

-> Player
Jogador Pacman

Função que encontra o Pacman.

changePS

Arguments

:: [Player]
Lista de todos os jogadores

-> Player
Jogador novo

-> [Player]
Lista de todos os jogadores com o novo jogador já inserido

Função que altera a lista de jogadores de modo a inserir o jogador que foi alterado.

mazePIc

Arguments

:: Maze
Labirinto

-> Int
Variável auxiliar

Função que verifica se o comprimento do labirinto é par ou ímpar, devolve um inteiro.

0, par

1, ímpar

mazePIa

Arguments

:: Maze
Labirinto

-> Int
Variável auxiliar

Função que verifica se a altura do labirinto é par ou ímpar, devolve um inteiro.

0, par

1, ímpar

houseExit

Arguments

:: Maze
Labirinto

-> Coords
Coordenada inicial da Casa dos Ghosts

-> Coords
Coordenada da saída

Função que descobre a coordenada da saída da Casa dos Ghosts.

Mudança de estados dos Ghosts

mainGhost

Arguments

:: Player
Jogador novo

-> Int
Pontuação anterior

-> Int
Pontuação nova

-> State
Estado novo

-> Coords
Coordenadas do jogador

-> State
Estado com a alteração dos estados dos Ghosts

Função que chama as respetivas funções dependendo das seguintes situações:

diferença entre a pontuação anterior e a nova == 5 que significa que o jogador comeu um Food Big

diferença entre a pontuação anterior e a nova == 11 || == 10 que significa que o jogador comeu um Ghost e um Food Little ou apenas um Ghost

diferença entre a pontuação anterior e a nova >= 15 que significa todos os outros casos possíveis, que vão ser analisados na respetiva função

changeGhostState

Arguments

:: Player
Jogador

-> Maze
Labirinto

-> [Player]
Lista de todos os jogadores

-> Coords
Coordenada do jogador

-> Int
Varíavel recebida

-> [Player]
Lista de jogadores com os Ghosts alterados

Função que chama as respetivas funções dependendo da situação do Ghost e da variável recebida.

changeGhostStateALL

Arguments

:: Player
Ĵogador

-> Maze
Labirinto

-> [Player]
Lista de todos os jogadores

-> Coords
Coordenada do jogador

-> [Player]
Lista de jogadores com os Ghosts alterados

Função que altera o estado de todos os Ghosts consoante, as várias condições propostas:

se Ghost está na posição certa e se jogador está em modo MEGA ou NORMAL

se jogador está em modo MEGA apenas

changeGhostStateAUX

Arguments

:: Maze
Labirinto

-> Player
Ghost pretendido

-> Int
Variável recebida

-> Player
Ghost novo

Função que altera o estado de um único Ghost, consoante a variável recebida.

middleCoordMaze

Arguments

:: Maze
Labirinto

-> Coords
Coordenadas do meio do labirinto

Função que encontra as coordenadas, na casa dos fantasmas, para a qual o Ghost se move quando morre.

heightMaze

Arguments

:: Maze
Labirinto

-> Int
Altura do labirinto

Função que determina a altura do labirinto.

lengthMaze

Arguments

:: Corridor
Corredor

-> Int
Largura do labirinto

Função que determina a largura do labirinto.

checkPacmanTM

Arguments

:: Player
Jogador

-> [Player]
Lista de todos os jogadores

-> [Player]
Lista de todos os jogadores com, possivelmente, os estados alterados

Função que verifica se o modo do pacman se encontra em Mega. Chama, depois, a função pretendida dependendo do resultado.

ghostTo

Arguments

:: [Player]
Lista de todos os jogadores

-> Int
Variável que decide o estado desejado

-> [Player]
Lista de todos os jogadores já com os estados alterados

Função que muda todos os Ghosts para o estado desejado.

ghostToAUX

Arguments

:: Player
Ghost

-> Int
Variável que decide o estado desejado

-> Player
Ghost novo

Função que muda um Ghost para o estado desejado dependendo da variável escolhida:

1: ghostToDead

2: ghostToAlive

Movimentação de Pacman ou Ghost

mainMove

Arguments

:: [Player]
Lista de todos os jogadores

-> Maze
Labirinto

-> Player
Jogador

-> Orientation
Orientação da jogada

-> Player
Jogador com o seu estado atualizado

Função que testa se o jogador está em modo Dying e se a orientação da jogada pretendida é igual à sua orientação. Chama, depois, as respetivas funções.

nextCoord

Arguments

:: [Player]
Lista de todos os jogadores

-> Player
Jogador

-> Maze
Labirinto

-> Player
Jogador com o seu estado atualizado

Função que determina a coordenada para a qual o jogador se quer mover. Chama, também, a função certa quando essa mesma coordenada. Passamos as coordenadas para coordenadas de labirinto, ou seja, incrementamos 1 em cada uma das coordenadas do tuplo.

nextMazeCoord

Arguments

:: [Player]
Lista de todos os jogadores

-> Player
Jogador

-> Maze
Labirinto

-> Coords
Coordenadas da jogada

-> Player
Jogador com o seu estado atualizado

Função que atualiza o estado do jogador para todas as seguintes situações:

Se a coordenada pretendida tiver um Food Little ou Food Big ou Empty então vai chamar a função para verificar a situação dos Ghosts nessa mesma coordenada

Se a coordenada pretendida tiver na posição de um túnel, vai chamar a função respetiva

mazeShowX

Arguments

:: Maze
Labirinto

-> Coords
Coordenada da peça pretendida

-> Piece
Peça pretendida

Função que percorre os corredores do labirinto até chegar ao corredor certo da coordenada pretendida. Chama, depois, a função que acaba a determinação desta peça.

mazeShowY

Arguments

:: Corridor
Corredor certo

-> Int
Largura restante até à coordenada

-> Piece
Peça pretendida

Função que determina a peça presente numa certa coordenada.

mazeShowT

Arguments

:: Maze
Labirinto

-> Coords
Coordenadas da jogada

-> Bool
Booleano pretendido

Função que testa se as coordenadas que nos queremos mover se encontram num túnel.

mazeTunnel

Arguments

:: [Player]
Lista de todos os jogadores

-> Player
Jogador

-> Maze
Labirinto

-> Coords
Coordenadas da jogada

-> Int
Pontuação acrescentar

-> Player
Jogador com o seu estado atualizado

Função que determina a nova coordenada do jogador quando queremos passar o túnel. Chama, depois, a função para verificar a situação dos Ghosts nessa mesma coordenada.

mazeGhost

Arguments

:: Player
Jogador

-> Coords
Coordenada pretendida

-> [Player]
Lista de todos os Ghost

-> Int
Pontuação

-> Player
Jogador

Função que testa se existe um Ghost nas coordenadas que o jogador se quer mover. Se sim, age conforme os modos do jogador e do Ghost:

Se em modo MEGA e o Ghost em Dead então incrementamos 10 na pontuação

Se em modo MEGA ou NORMAL e o Ghost em Alive então a função que verifica as vidas do jogador é chamada

mazeEatPacman

Arguments

:: Player
Jogador Pacman anterior

-> Player
Jogador Pacman novo

-> Player
Jogador Ghost anterior

-> Player
Jogador Ghost novo

-> Player
Jogador Pacman

Função que modifica o jogador dependendo da posição do fantasma. Função só chamada quando a jogada é feita por um fantasma.

lifeChanger

Arguments

:: Player
Jogador

-> Int
Variável auxiliar

-> Player
Jogador com o estado novo

Função que altera o modo do jogador para Dying caso não tenha mais nenhuma vida.

changeMouth

Arguments

:: Player
Jogador

-> Player
Jogador com boca alterada

Função que muda a boca do pacman

changeMegaTM

Arguments

:: Player
Jogador

-> Player
Jogador com tempo mega alterado

Função que diminui o tempo mega em 250ms

getGhosts

Arguments

:: [Player]
Lista de todos os jogadores

-> [Player]
Lista de todos os Ghosts

Função que determina a lista de todos os Ghosts.

getPacman :: [Player] -> Player

getGhostMode

Arguments

:: Player
Ghost

-> GhostMode
Modo do Ghost

Função que determina o modo em que um Ghost está.

Funções que correspondem ao labirinto

foodToEmpty

Arguments

:: Maze
Labirinto

-> Player
Jogador

-> Coords
Coordenadas pretendidas

-> Maze
Labirinto com a coordenada alterada

Função que altera no labirinto a comida que o Pacman comeu por Empty.

isThereFood

Arguments

:: Maze
Labirinto

-> Bool
Booleano pretendido

Função que invoca a função pretendida para percorrer o labirinto, corredor a corredor.

isThereFoodAUX

Arguments

:: Corridor
Corredor do labirinto

-> Bool
Booleano pretendido

Função que percorre o corredor pretendido para verificar se tem alguma comida no labirinto.

True significa que tem comida no labirinto

False significa que não tem nenhuma comida no labirinto