Tarefa3

Description

O objetivo deste módulo é a compactação do labirinto, facilitando a leitura do mesmo.

Compactação do labirinto

compactMaze

Arguments

:: Maze
Labirinto

-> Instructions
Instruções da compactação do labirinto

Compactação por corredor

compactCorridor

Arguments

:: Maze
Labirinto

-> [[(Int, Piece)]]
Matriz de tuplos com a compactação por corredor

Função que percorre o labirinto e cria uma matriz com todos os corredores compactados.

compactCorridorAUX

Arguments

:: Corridor
Corredor

-> Int
Variável que contém o número de peças para passar à frente no corredor, ou seja, que já estão compactadas

-> [(Int, Piece)]
Lista de tuplos com a compactação do corredor

Função que compacta um corredor numa lista de tuplos.

rowInt

Arguments

:: Corridor
Corredor

-> Piece
Peça

-> Int
Variável que guarda o número de peças seguidas iguais

-> Int
Número de peças seguidas iguais

Função que determina o número de peças seguidas iguais à peça recebida.

Aplicação das Instructions

compactRepeat

Arguments

:: [[(Int, Piece)]]
Matriz de tuplos com a compactação por corredor

-> [[(Int, Piece)]]
Matriz de tuplos com a compactação por corredor

-> Int
Variável que incrementa sempre que esta função é chamada, representa o número do corredor atual

-> Instructions
Instruções da compactação do labirinto

Função que percorre a matriz e para cada lista de tuplos chama a função para aplicar as instruções.

compactRepeatAUX

Arguments

:: [[(Int, Piece)]]
Matriz de tuplos com a compactação por corredor

-> [(Int, Piece)]
Lista de tuplos com a compactação do corredor

-> Int
Variável que incrementa sempre que esta função é chamada, representa o número do corredor para repetir

-> Int
Variável que decrementa sempre que esta função é chamada, representa o número do corredor atual até ao corredor entregue

-> Int
Variável que guarda o número do corredor entregue

-> Instruction
Instrução da compactação do corredor

Função que percorre a matriz e para cada lista de tuplos compara com a lista de tuplos entregue.

Se igual, então insere a Instruction Repeat com o número desse corredor

Se diferente, então volta a repetir a função até chegar ao próprio corredor inserindo depois a Instruction Instruct