Tarefa4

Description

Introdução

Nesta tarefa tivemos de fazer jogar os jogadores (Pacman e Ghosts) consoante a passagem de tempo (Steps).

Objetivos

O objetivo principal era fazer jogar os jogadores, jogando primeiro o Pacman e de seguida os Ghosts. Para isto usamos a Tarefa2 já anteriormente desenvolvida, usando primeiramente a função play para o Pacman e de seguida para todos os Ghosts. O movimento do Pacman pode funcionar de duas formas, através dos inputs do jogador ou através de um bot desenvolvido na Tarefa6. Criamos então uma função movePacmanBot que ao invés de utilizar a função play da Tarefa2 com a orientação do Pacman dada pelo utilizador, vai a Tarefa6 buscar as jogadas que o bot vai realizar e então chamar a função play da Tarefa2 para essas jogadas. Um outro objetivo era a introdução de diferentes velocidades possíveis para os Ghosts, de modo a podermos ter um fator de dificuldade, aumentando a velocidade dos mesmos a cada nível.

Discussão e Conclusão

Acreditamos que a tarefa cumpre com o seu principal objetivo, fazer jogar os jogadores e a opção de um bot para mover o Pacman, falhando apenas no facto referido acima das velocidades, que após varias tentativas, não foi possivel implementar e, portanto, todos os jogadores se mexem à mesma velocidade.

Documentation

defaultDelayTime :: Integer

Funções principais para a realização de todas as jogadas de todas as maneiras

passTime

Arguments

:: Int
Steps

-> State
Estado atual

-> Int
Modo de jogo

-> Int
Pausa

-> State
Estado novo

Função que consoante a passagem de tempo, faz jogar todos os jogadores.

movePacman

Arguments

:: State
Estado atual

-> State
Estado novo

Função que realiza uma jogada do Pacman.

moveGhosts

Arguments

:: [Play]
Lista de todas as jogadas dos Ghost

-> State
Estado atual

-> State
Estado novo

Função que permite realizar todas as jogadas pretendidas do bot dos Ghost.

movePacmanBot

Arguments

:: Maybe Play
Jogada do Pacman

-> State
Estado atual

-> State
Estado novo

Função que permite realizar a jogada pretendida do bot do Pacman.