Tarefa5

Description

Introdução

Nesta tarefa prentende-se criar as jogadas possiveis para os Ghosts, em dois modos diferentes, Chase e Scatter.

Objetivos

O objetivo deste módulo é a realização das jogadas automáticas dos Ghosts, consoante todas as diferentes situações em que um determinado Ghost se encontra e o estado do jogo. Por exemplo, se um Ghost esta em Chase mode, as jogadas retornadas pela função ghostPlay serão jogadas que aproximam o Ghost do Pacman. O inverso acontece quando estamos em Scatter mode". No Chase mode para a decisão da jogada do Ghost foram definidos os seguintes passos, em ordem de utilização: 1 -> Pacman na casa seguinte ao Ghost para qualquer orientação; 2 -> Posição do Ghost em relação à posição do Pacman. 1: Se isto for verdade, então o Ghost movimenta-se para essa posição. 2: É comparada a posição atual do Pacman com a posição atual do Ghost e é definida a melhor orientação para a movimentação do mesmo. Por exemplo, se o Pacman estiver à esquerda e em baixo o Ghost vai tentar se deslocar nessa direção, tendo em conta qual das duas orientações o Pacman está mais perto. Ou seja, comparando a distância entre as duas coordenadas de cada um. Depois de definida a melhor orientação para o movimento, é dada uma lista de critérios para ver se esse movimento é possível ou não. Sendo essa lista a seguinte, por ordem de prioridade: Orientação pretendida não vai contra uma parede; (*)Nova orientação escolhida não é contrária à orientação do Pacman em conjunto com a possibilidade de se deslocar para a orientação anteriormente escolhida e não vai contra uma parede; (*)Nova orientação escolhida não vai contra uma parede.

(*)Esta nova orientação é definida sempre por esta ordem, Esquerda - Direita - Baixo - Cima

Já no Scatter mode ele faz exatamente o contrário da Chase mode com os mesmos critérios.

Tendo em conta que quando a orientação é para Baixo ele nunca consegue entrar dentro da Casa dos Ghosts. Sempre que um Ghost está dentro da Casa dos Ghosts ele como prioridade é sair da casa sempre. A movimentação para os túneis está feita, mas infelizmente não está a funcionar muito bem, por isso ele nem sempre se movimenta para lá.

Discussão e Conclusão

Entendemos que esta função, apesar de não o fazer da melhor forma possivel, cria jogadas possiveis e de acordo com o modo em que o ghost se encontra, e, por isso, cumpre com os seus objetivos.

Realização das jogadas dos Ghosts

ghostPlay

Arguments

:: State
Estado do jogo

-> [Play]
Lista das jogadas de todos os Ghosts

Função principal que invoca todas as outras para obter todas as jogadas dos Ghosts.

modeDecider

Arguments

:: State
Estado do jogo

-> Player
Pacman

-> [Play]
Lista das jogadas de todos os Ghosts

Função que, dependendo do modo do Ghost, invoca a função do modo pretendido.

Funções que decidem a jogada para o modo Chase

chaseMode

Arguments

:: Maze
Labirinto

-> Player
Ghost

-> Player
Pacman

-> Play
Jogada do Ghost

Função que determina o movimento do Ghost se o Pacman estiver numa posição a uma distância de 1.

chaseModeAUX

Arguments

:: Maze
Labirinto

-> Player
Ghost

-> Player
Pacman

-> Play
Jogada do Ghost

Função que decide qual a orientação para a jogada do Ghost dependendo da posição atual do Pacman.

Funções que decidem a jogada para o modo Scatter

scatterMode

Arguments

:: Maze
Labirinto

-> Player
Ghost

-> Player
Pacman

-> Play
Jogada do Ghost

Função que determina o movimento do Ghost se o Pacman estiver numa posição a uma distância de 1.

scatterModeAUX

Arguments

:: Maze
Labirinto

-> Player
Ghost

-> Player
Pacman

-> Play
Jogada do Ghost

Função que decide qual a orientação para a jogada do Ghost dependendo da posição atual do Pacman.

oppMove

Arguments

:: Maze
Labirinto

-> Player
Ghost

-> Player
Pacman

-> Play
Jogada do Ghost

Função que decide qual a orientação para a jogada do Ghost dependendo da orientação do Pacman.

Realização da jogada do Ghost

mLeft

Arguments

:: Maze
Labirinto

-> Player
Ghost

-> Player
Pacman

-> Orientation
Orientação auxiliar

-> Play
Jogada do Ghost

Função que verifica se o Ghost se encontra numa posição perto do túnel, ao se mover para a esquerda, invocando as devidas funções para cada caso.

mRight

Arguments

:: Maze
Labirinto

-> Player
Ghost

-> Player
Pacman

-> Orientation
Orientação auxiliar

-> Play
Jogada do Ghost

Função que verifica se o Ghost se encontra numa posição perto do túnel, ao se mover para a direta, invocando as devidas funções para cada caso.

mDown

Arguments

:: Maze
Labirinto

-> Player
Ghost

-> Player
Pacman

-> Orientation
Orientação auxiliar

-> Play
Jogada do Ghost Mover só para baixo

Função que realiza a jogada do Ghost com prioridade para a movimentação para baixo.

mUp

Arguments

:: Maze
Labirinto

-> Player
Ghost

-> Player
Pacman

-> Orientation
Orientação auxiliar

-> Play
Jogada do Ghost Mover só para cima

Função que realiza a jogada do Ghost com prioridade para a movimentação para cima.

mDownLeft

Arguments

:: Maze
Labirinto

-> Player
Ghost

-> Player
Pacman

-> Play
Jogada do Ghost

Função que verifica qual a melhor orientação possível para a jogada do Ghost. Definindo a prioridade em movimentação para baixo ou para a esquerda.

mDownRight

Arguments

:: Maze
Labirinto

-> Player
Ghost

-> Player
Pacman

-> Play
Jogada do Ghost

Função que verifica qual a melhor orientação possível para a jogada do Ghost. Definindo a prioridade em movimentação para baixo ou para a direita.

mUpLeft

Arguments

:: Maze
Labirinto

-> Player
Ghost

-> Player
Pacman

-> Play
Jogada do Ghost

Função que verifica qual a melhor orientação possível para a jogada do Ghost. Definindo a prioridade em movimentação para cima ou para a esquerda.

mUpRight

Arguments

:: Maze
Labirinto

-> Player
Ghost

-> Player
Pacman

-> Play
Jogada do Ghost

Função que verifica qual a melhor orientação possível para a jogada do Ghost. Definindo a prioridade em movimentação para cima ou para a direita.

Funções auxiliares para a realização da jogada do Ghost

mTunnel

Arguments

:: Maze
Labirinto

-> Player
Ghost

-> Player
Pacman

-> Orientation
Orientação pretendida para o movimento

-> Orientation
Orientação auxiliar

-> Play
Jogada do Ghost

Função que verifica e realiza a movimentação do Ghost se estiver numa posição que consegue aceder ao túnel, verificando também se vale ou não a pena esse movimento.

mTunnelAUX

Arguments

:: Maze
Labirinto

-> Player
Ghost

-> Player
Pacman

-> Orientation
Orientação pretendida para o movimento

-> Orientation
Orientação auxiliar

-> Play
Jogada do Ghost

Função que verifica se vale ou não a pena movimentar para o túnel, invocando a devida função caso não.

mAUX

Arguments

:: Maze
Labirinto

-> Player
Ghost

-> Coords
Coordenadas do Pacman

-> Orientation
Orientação do Pacman

-> Orientation
Orientação pretendida para o movimento

-> Orientation
Orientação auxiliar

-> Play
Jogada do Ghost Mover só para a esquerda

Função que realiza a jogada do Ghost dependendo das prioridades de orientação dadas.

Funções auxiliares para a movimentação para fora da Casa dos Ghosts

insideHouse

Arguments

:: Coords
Coordenada inicial da Casa dos Ghosts

-> Coords
Coordenada final da Casa dos Ghosts

-> Coords
Coordenada do Ghost

-> Bool
Verificação se está dentro ou não

Função que deteta se o Ghost está dentro da Casa dos Ghosts.

finalCoordCF

Arguments

:: Maze
Labirinto

-> Coords
Coordenada inicial da Casa dos Ghosts

-> Coords
Coordenada final da Casa dos Ghosts

Função que devolve a posição final da Casa dos Ghosts.

mExit

Arguments

:: Maze
Labirinto

-> Player
Ghost

-> Player
Pacman

-> Play
Jogada do Ghost

Função que determina o movimento ideal para o Ghost sair da Casa dos Ghosts.