Tarefa6

Description

Introdução

Esta tarefa consiste na criação de um bot capaz de jogar sem input do utilizador.

Objetivos

O objetivo deste módulo é a realização das jogadas automáticas do Pacman, consoante todas as diferentes situações em que um determinado Pacman se encontra e o estado do jogo. As jogadas fazem com que o pacman coma comidas, mas tendo em consideração a presença de Ghosts por perto. A estratégia utilizada tem sempre como prioridade a movimentação para a Food Big e depois sim para um Food Little. Primeiramente o Pacman vê se numa jogada com a sua orientação tem qualquer tipo de Food, se sim vai pra lá. Se não, ele vai percorrer todas as outras orientações e procurar por Food. Se mesmo assim, não conseguir encontrar vai, novamente, percorrer todas as orientações, mas desta vez com uma distância de 2 jogadas. Por alguma razão, se não encontrar nenhuma comida, ele tenta apenas se movimentar para alguma posição diferente de uma parede. Depois de definida a coordenada para onde se quer deslocar, é considerada as posições atuais dos Ghosts e se tiver um na posição onde ele quer ir, volta a procurar outra orientação.

Tendo sempre em conta que o Pacman também nunca consegue entrar na Casa dos Ghosts. Foi considerado também o movimento para o túnel, de forma a distanciar-se do Ghost, que cumpre perfeitamente.

Discussão e Conclusão

Entendemos que esta função, apesar de não o fazer da melhor forma possivel, cria jogadas possiveis para o pacman tendo em conta varias situações, como a necessidade de comer e a presença de Ghosts, e, por isso, cumpre com os seus objetivos.

Realização das jogadas do Pacman

bot

Arguments

:: Int
ID do Pacman

-> State
Estado do jogo

-> Maybe Play
Jogada do Pacman

Função principal que invoca todas as outras para obter a melhor jogada do Pacman.

bMove

Arguments

:: Maze
Labirinto

-> [Player]
Lista de todos os Ghost

-> Player
Jogador Pacman

-> Maybe Play
Jogada do Pacman

Função que realiza a melhor jogada possível do Pacman.

bMoveL

Arguments

:: Maze
Labirinto

-> [Player]
Lista de todos os Ghost

-> Player
Jogador Pacman

-> Maybe Play
Jogada do Pacman

Função que decide a melhor jogada tendo em conta a orientação para a esquerda.

bMoveR

Arguments

:: Maze
Labirinto

-> [Player]
Lista de todos os Ghost

-> Player
Jogador Pacman

-> Maybe Play
Jogada do Pacman

Função que decide a melhor jogada tendo em conta a orientação para a direita.

bMoveU

Arguments

:: Maze
Labirinto

-> [Player]
Lista de todos os Ghost

-> Player
Jogador Pacman

-> Maybe Play
Jogada do Pacman

Função que decide a melhor jogada tendo em conta a orientação para cima.

bMoveD

Arguments

:: Maze
Labirinto

-> [Player]
Lista de todos os Ghost

-> Player
Jogador Pacman

-> Maybe Play
Jogada do Pacman

Função que decide a melhor jogada tendo em conta a orientação para baixo.

Funções auxiliares para a concretização da jogada do Pacman

bOrientation

Arguments

:: Maze
Labirinto

-> [Player]
Lista de todos os Ghost

-> Player
Jogador Pacman

-> Orientation
Orientação não pretendida

-> Int
Variável auxiliar

-> Maybe Play
Jogada do Pacman

Função que decide qual a nova melhor orientação para o Pacman.

bOrientationAUX

Arguments

:: Maze
Labirinto

-> [Player]
Lista de todos os Ghost

-> Player
Jogador Pacman

-> Orientation
Orientação não pretendida

-> Int
Variável auxiliar

-> Maybe Play
Jogada do Pacman

Função que decide qual a nova melhor orientação para o Pacman, tendo em conta que vai ser necessário dois movimentos.

bFinalMove

Arguments

:: Maze
Labirinto

-> [Player]
Lista de todos os Ghost

-> Player
Jogador Pacman

-> Int
Variável auxiliar

-> Maybe Play
Jogada do Pacman

Função que realiza a última tentativa de movimentação para o Pacman.

Função para a realização do movimento do Pacman para o túnel

bTunnel

Arguments

:: Maze
Labirinto

-> [Player]
Lista de todos os Ghost

-> Player
Jogador Pacman

-> Orientation
Lado do Labirinto que o Pacman está

-> Orientation
Orientação do Pacman

-> Int
Variável auxiliar

-> Maybe Play
Jogada do Pacman Lado esquerdo do labirinto

Função que realiza o movimento do pacman para o túnel.

Verificação da posição dos Ghosts

bGhosts

Arguments

:: Maze
Labirinto

-> [Player]
Lista de todos os Ghost usada para a seguinte invocação

-> [Player]
Lista de todos os Ghost

-> Player
Jogador Pacman

-> Coords
Coordenada pretendida

-> Orientation
Orientação pretendida

-> Int
Variável auxiliar

-> Maybe Play
Jogada do Pacman

Função que verifica se existe um Ghost na posição que o Pacman quer ir.