TestesT2

Documentation

maze1 :: [[Piece]]

maze2 :: [[Piece]]

maze3 :: [[Piece]]

player0 :: Player

player1 :: Player

player2 :: Player

player3 :: Player

player4 :: Player

player5 :: Player

player6 :: Player

playerList :: [Player]

playerList2 :: [Player]

estado1 :: State

estado2 :: State

estado3 :: State

estado4 :: State

testCases :: [([Play], State)]

Testes da Tarefa2 para os casos em baixo indicados

testCases1 :: [([Play], State)]

Testes movimento para todas as direções com o pacman com orientação diferente da direção

testCases2 :: [([Play], State)]

Testes movimento para todas as direções com o pacman orientado na direção

testCases3 :: [([Play], State)]

Testes de passar os tuneis e passar os tuneis e voltar

testCases4 :: [([Play], State)]

Testes comida pequena, grande, empty

testCases5 :: [([Play], State)]

Testes Comer 1 fantasma e empty, comer 1 fantasma e comida grande, comer 1 fantasma e comida pequena

testCases6 :: [([Play], State)]

Testes comer 2 ou mais fantasmas e estes voltarem para o centro

testCases7 :: [([Play], State)]

Testes colisao com fantasmas e vida = 1 e nao mexer quando esta Dying

testCases8 :: [([Play], State)]

Testes colisao com fantasmas e vida > 1